+++ app/controllers/homebrews_v2/feats_controller.rb
@@ -5,18 +5,20 @@ module HomebrewsV2
     include SerializeRelation
     include SerializeResource
 
+    before_action :find_feats, only: %i[index copy]
+    before_action :find_upvotes, only: %i[index]
     before_action :find_feat, only: %i[show]
     before_action :find_another_feat, only: %i[copy]
     before_action :find_feats_for_batch_destroy, only: %i[batch_destroy]
 
     def index
       serialize_relation_v2(
-        feats,
+        @feats,
         serializer,
         :homebrews,
-        serialized_fields: { only: %i[id title own books] },
+        serialized_fields: { only: %i[id title own books upvotes_count upvoted] },
         order_options: order_options,
-        context: { current_user_id: current_user.id }
+        context: { current_user_id: current_user.id, upvotes: @upvotes }
       )
     end
 
@@ -41,12 +43,16 @@ module HomebrewsV2
 
     private
 
+    def find_upvotes
+      @upvotes = current_user.upvotes.where(upvoteable_type: 'Feat').pluck(:upvoteable_id)
+    end
+
     def find_feat
       @feat = class_name.find(params.expect(:id))
     end
 
     def find_another_feat
-      @feat = feats.find(params.expect(:id))
+      @feat = @feats.find(params.expect(:id))
     end
 
     def find_feats_for_batch_destroy