+++ app/models/upvote.rb
@@ -2,5 +2,5 @@
 
 class Upvote < ApplicationRecord
   belongs_to :user
-  belongs_to :upvoteable, polymorphic: true
+  belongs_to :upvoteable, polymorphic: true, counter_cache: :upvotes_count
 end