+++ app/platforms/daggerheart/homebrews/mechanic_item.rb
@@ -11,15 +11,17 @@ module Daggerheart
     class MechanicItem < ::Homebrew
       attribute :info, Daggerheart::Homebrews::MechanicItemData.to_type
 
-      def to_homebrew_json
+      def to_homebrew_json(with_id: true)
         [
           {
-            id: id,
+            id: with_id ? id : nil,
             title: title,
             description: description,
             tier: info.tier,
-            features: Daggerheart::Feat.where(origin: 'mechanic', origin_value: id).map(&:to_homebrew_json)
-          }
+            features: Daggerheart::Feat.where(origin: 'mechanic', origin_value: id).map { |item|
+              item.to_homebrew_json(with_id: with_id)
+            }
+          }.compact
         ]
       end
     end