+++ app/platforms/daggerheart/homebrews/subclass.rb
@@ -13,18 +13,20 @@ module Daggerheart
     class Subclass < ::Homebrew
       attribute :info, Daggerheart::Homebrews::SubclassData.to_type
 
-      def to_homebrew_json
+      def to_homebrew_json(with_id: true)
         [
           {
-            id: id,
+            id: with_id ? id : nil,
             title: title,
             description: description,
             public: attributes['public'],
             class_id: info.class_id,
             spellcast: info.spellcast,
             mechanics: info.mechanics,
-            features: Daggerheart::Feat.where(origin_value: id).map(&:to_homebrew_json)
-          }
+            features: Daggerheart::Feat.where(origin_value: id).map { |item|
+              item.to_homebrew_json(with_id: with_id)
+            }
+          }.compact
         ]
       end
     end