+++ app/serializers/homebrews_v2/daggerheart/domain_serializer.rb
@@ -10,10 +10,10 @@ module HomebrewsV2
         return [] unless context[:features]
 
         relation = context[:features]
-        Panko::ArraySerializer.new(
-          relation,
-          each_serializer: HomebrewsV2::Daggerheart::FeatSerializer
-        ).serialize(relation)
+        Panko::ArraySerializer
+          .new(relation, each_serializer: HomebrewsV2::Daggerheart::FeatSerializer)
+          .serialize(relation)
+          .sort_by! { |item| item.dig('conditions', 'level') }
       end
     end
   end