+++ app/serializers/homebrews_v2/list_item_serializer.rb
@@ -2,7 +2,7 @@
 
 module HomebrewsV2
   class ListItemSerializer < ApplicationSerializer
-    attributes :id, :title, :description, :own, :books
+    attributes :id, :title, :description, :own, :books, :upvotes_count, :upvoted
 
     def title
       translate(object.name)
@@ -17,6 +17,13 @@ module HomebrewsV2
       return false unless context[:current_user_id]
 
       object.user_id == context[:current_user_id]
+    end
+
+    def upvoted # rubocop: disable Naming/PredicateMethod
+      return false unless context
+      return false unless context[:upvotes]
+
+      context[:upvotes].include?(object.id)
     end
 
     def books