+++ spec/controllers/homebrews_v2/daggerheart/books_controller_spec.rb
@@ -21,7 +21,7 @@ describe HomebrewsV2::Daggerheart::BooksController do
         expect(response).to have_http_status :ok
         expect(response.parsed_body['homebrews'].size).to eq 2
         expect(response.parsed_body.dig('homebrews', 0).keys).to(
-          contain_exactly('id', 'title', 'provider', 'shared', 'public', 'enabled', 'own')
+          contain_exactly('id', 'title', 'provider', 'shared', 'public', 'enabled', 'own', 'upvoted', 'upvotes_count')
         )
       end
     end